fixed relative jumping

+++ cpu/cu.c
@@ -13,7 +13,7 @@ void inc(uint16_t *program_counter)
 /*
  *
  */
-void jmp(uint16_t* program_counter, uint16_t offset)
+void jmp(uint16_t* program_counter, int8_t offset)
 {
     *program_counter += offset;
 }
