+++ cpu/cu.h
@@ -12,6 +12,6 @@ void inc(uint16_t* program_counter);
 /*
  * icnrements the programm counter with offset
  */
-void jmp(uint16_t* program_counter, uint16_t offset);
+void jmp(uint16_t* program_counter, int8_t offset);
 
 #endif
